+++ src/components/admin/routes/subCategories/SubCategoryCardActions.tsx
@@ -14,6 +14,9 @@ import {
 } from "@/components/shared/CustomToast";
 import LoadingSpinner from "@/components/shared/LoadingSpinner";
 import Modal from "@/components/shared/Modal";
+import { Permission } from "@/enums/permission.enum";
+import { usePermission } from "@/hooks/usePermission";
+import { showNoPermissionToast } from "@/utils/permissionToast";
 
 type DashboardCardActionsProps<
   T extends { _id: string; isDeleted: boolean; isActive: boolean },
@@ -49,7 +52,25 @@ const CategoryCardActions = <
   const [isLoading, setIsLoading] = useState(false);
   const [isEditModalOpen, setIsEditModalOpen] = useState(false);
 
+  const {
+    canDeleteSubCategory,
+    canRestoreSubCategory,
+    canActivateSubCategory,
+    canDeActivateSubCategory,
+    canUpdateSubCategory,
+  } = usePermission({
+    canDeleteSubCategory: Permission.SUB_CATEGORIES_DELETE,
+    canRestoreSubCategory: Permission.SUB_CATEGORIES_RESTORE,
+    canActivateSubCategory: Permission.SUB_CATEGORIES_ACTIVATE,
+    canDeActivateSubCategory: Permission.SUB_CATEGORIES_DEACTIVATE,
+    canUpdateSubCategory: Permission.SUB_CATEGORIES_UPDATE,
+  });
+
   const handleDelete = useCallback(async () => {
+    if (!canDeleteSubCategory) {
+      showNoPermissionToast(t);
+      return;
+    }
     setIsLoading(true);
     try {
       const resp = await deleteFn(cardItem._id);
@@ -70,9 +91,13 @@ const CategoryCardActions = <
       setIsLoading(false);
       await invalidateQuery(queryClient, queryKey);
     }
-  }, [deleteFn, cardItem._id, queryClient, queryKey, t]);
+  }, [deleteFn, cardItem._id, queryClient, queryKey, t, canDeleteSubCategory]);
 
   const handleUnDelete = useCallback(async () => {
+    if (!canRestoreSubCategory) {
+      showNoPermissionToast(t);
+      return;
+    }
     setIsLoading(true);
     try {
       const resp = await unDeleteFn(cardItem._id);
@@ -93,9 +118,25 @@ const CategoryCardActions = <
       setIsLoading(false);
       await invalidateQuery(queryClient, queryKey);
     }
-  }, [unDeleteFn, cardItem._id, queryClient, queryKey, t]);
+  }, [
+    unDeleteFn,
+    cardItem._id,
+    queryClient,
+    queryKey,
+    t,
+    canRestoreSubCategory,
+  ]);
 
   const handleToggleActiveStatus = useCallback(async () => {
+    const hasAccess = cardItem?.isActive
+      ? canDeActivateSubCategory
+      : canActivateSubCategory;
+
+    if (!hasAccess) {
+      showNoPermissionToast(t);
+      return;
+    }
+
     setIsLoading(true);
     try {
       const resp = await switchUserActiveStatusFn(
@@ -129,6 +170,8 @@ const CategoryCardActions = <
     queryKey,
     locale,
     t,
+    canDeActivateSubCategory,
+    canActivateSubCategory,
   ]);
 
   return (
@@ -147,43 +190,47 @@ const CategoryCardActions = <
         )}
 
         <div className="w-1/4">
-          <ToggleSwitch
-            value={cardItem.isActive}
-            onChange={handleToggleActiveStatus}
-            width={50}
-            height={22}
-            trackColorInactive="#E55050"
-            trackColorActive="#16610E"
-            isDisabled={false}
-          />
+          {(canDeActivateSubCategory || canActivateSubCategory) && (
+            <ToggleSwitch
+              value={cardItem.isActive}
+              onChange={handleToggleActiveStatus}
+              width={50}
+              height={22}
+              trackColorInactive="#E55050"
+              trackColorActive="#16610E"
+              isDisabled={false}
+            />
+          )}
         </div>
 
         <div className="w-full flex items-center justify-center gap-4 flex-wrap sm:flex-nowrap">
-          {!cardItem.isDeleted ? (
-            <Button
-              disabled={isLoading}
-              className={`${
-                showEditButton ? "w-full" : "w-full"
-              } min-h-3 bg-red-500 hover:bg-red-600 text-white-50 transition-all`}
-              onClick={handleDelete}
-            >
-              <Package className="w-1 h-1" />
-              {t("general.actions.delete")}
-            </Button>
-          ) : (
-            <Button
-              disabled={isLoading}
-              className={`${
-                showEditButton ? "w-full" : "w-full"
-              } min-h-3 bg-success-500 hover:bg-success-600 text-white-50 transition-all`}
-              onClick={handleUnDelete}
-            >
-              <PackageOpen className="w-1 h-1" />
-              {t("general.actions.restore")}
-            </Button>
-          )}
+          {!cardItem.isDeleted
+            ? canDeleteSubCategory && (
+                <Button
+                  disabled={isLoading}
+                  className={`${
+                    showEditButton ? "w-full" : "w-full"
+                  } min-h-3 bg-red-500 hover:bg-red-600 text-white-50 transition-all`}
+                  onClick={handleDelete}
+                >
+                  <Package className="w-1 h-1" />
+                  {t("general.actions.delete")}
+                </Button>
+              )
+            : canRestoreSubCategory && (
+                <Button
+                  disabled={isLoading}
+                  className={`${
+                    showEditButton ? "w-full" : "w-full"
+                  } min-h-3 bg-success-500 hover:bg-success-600 text-white-50 transition-all`}
+                  onClick={handleUnDelete}
+                >
+                  <PackageOpen className="w-1 h-1" />
+                  {t("general.actions.restore")}
+                </Button>
+              )}
 
-          {showEditButton && renderEditForm && (
+          {showEditButton && canUpdateSubCategory && renderEditForm && (
             <Button
               disabled={isLoading}
               className="w-full min-h-3 bg-gray-500 hover:bg-gray-600 text-white-50 transition-all"
@@ -196,7 +243,7 @@ const CategoryCardActions = <
         </div>
       </div>
 
-      {renderEditForm && (
+      {renderEditForm && canUpdateSubCategory && (
         <Modal
           isOpen={isEditModalOpen}
           onClose={() => setIsEditModalOpen(false)}
