+++ src/components/admin/routes/typeHintConfigs/TypeHintConfigCardActions.tsx
@@ -17,6 +17,9 @@ import {
 import LoadingSpinner from "@/components/shared/LoadingSpinner";
 import Modal from "@/components/shared/Modal";
 import { useHomeEffectsContext } from "@/contexts/HomeEffectsContext";
+import { Permission } from "@/enums/permission.enum";
+import { usePermission } from "@/hooks/usePermission";
+import { showNoPermissionToast } from "@/utils/permissionToast";
 
 type DashboardCardActionsProps<
   T extends { _id: string; isDeleted: boolean; isActive: boolean },
@@ -53,7 +56,26 @@ const TypeHintConfigCardActions = <
   const [isLoading, setIsLoading] = useState(false);
   const [isEditModalOpen, setIsEditModalOpen] = useState(false);
 
+  const {
+    canDeleteTypeHintConfig,
+    canRestoreTypeHintConfig,
+    canActivateTypeHintConfig,
+    canDeActivateTypeHintConfig,
+    canUpdateTypeHintConfig,
+  } = usePermission({
+    canDeleteTypeHintConfig: Permission.TYPE_HINT_CONFIGS_DELETE,
+    canRestoreTypeHintConfig: Permission.TYPE_HINT_CONFIGS_RESTORE,
+    canActivateTypeHintConfig: Permission.TYPE_HINT_CONFIGS_ACTIVATE,
+    canDeActivateTypeHintConfig: Permission.TYPE_HINT_CONFIGS_DEACTIVATE,
+    canUpdateTypeHintConfig: Permission.TYPE_HINT_CONFIGS_UPDATE,
+  });
+
   const handleDelete = useCallback(async () => {
+    if (!canDeleteTypeHintConfig) {
+      showNoPermissionToast(t);
+      return;
+    }
+
     setIsLoading(true);
     try {
       const resp = await deleteFn(cardItem._id, locale);
@@ -91,9 +113,14 @@ const TypeHintConfigCardActions = <
     ,
     setChangeBanners,
     t,
+    canDeleteTypeHintConfig,
   ]);
 
   const handleUnDelete = useCallback(async () => {
+    if (!canRestoreTypeHintConfig) {
+      showNoPermissionToast(t);
+      return;
+    }
     setIsLoading(true);
     try {
       const resp = await unDeleteFn(cardItem._id, locale);
@@ -131,9 +158,19 @@ const TypeHintConfigCardActions = <
     locale,
     setChangeBanners,
     t,
+    canRestoreTypeHintConfig,
   ]);
 
   const handleToggleActiveStatus = useCallback(async () => {
+    const hasAccess = cardItem?.isActive
+      ? canDeActivateTypeHintConfig
+      : canActivateTypeHintConfig;
+
+    if (!hasAccess) {
+      showNoPermissionToast(t);
+      return;
+    }
+
     setIsLoading(true);
     try {
       const resp = await switchUserActiveStatusFn(
@@ -175,6 +212,8 @@ const TypeHintConfigCardActions = <
     locale,
     setChangeBanners,
     t,
+    canDeActivateTypeHintConfig,
+    canActivateTypeHintConfig,
   ]);
 
   return (
@@ -193,43 +232,47 @@ const TypeHintConfigCardActions = <
         )}
 
         <div className="w-1/4">
-          <ToggleSwitch
-            value={cardItem.isActive}
-            onChange={handleToggleActiveStatus}
-            width={50}
-            height={22}
-            trackColorInactive="#E55050"
-            trackColorActive="#16610E"
-            isDisabled={false}
-          />
+          {(canActivateTypeHintConfig || canDeActivateTypeHintConfig) && (
+            <ToggleSwitch
+              value={cardItem.isActive}
+              onChange={handleToggleActiveStatus}
+              width={50}
+              height={22}
+              trackColorInactive="#E55050"
+              trackColorActive="#16610E"
+              isDisabled={false}
+            />
+          )}
         </div>
 
         <div className="w-full flex items-center justify-center gap-4 flex-wrap sm:flex-nowrap">
-          {!cardItem.isDeleted ? (
-            <Button
-              disabled={isLoading}
-              className={`${
-                showEditButton ? "w-full" : "w-full"
-              } min-h-3 bg-red-500 hover:bg-red-600 text-white-50 transition-all`}
-              onClick={handleDelete}
-            >
-              <Package className="w-1 h-1" />
-              {t("general.actions.delete")}
-            </Button>
-          ) : (
-            <Button
-              disabled={isLoading}
-              className={`${
-                showEditButton ? "w-full" : "w-full"
-              } min-h-3 bg-success-500 hover:bg-success-600 text-white-50 transition-all`}
-              onClick={handleUnDelete}
-            >
-              <PackageOpen className="w-1 h-1" />
-              {t("general.actions.restore")}
-            </Button>
-          )}
+          {!cardItem.isDeleted
+            ? canDeleteTypeHintConfig && (
+                <Button
+                  disabled={isLoading}
+                  className={`${
+                    showEditButton ? "w-full" : "w-full"
+                  } min-h-3 bg-red-500 hover:bg-red-600 text-white-50 transition-all`}
+                  onClick={handleDelete}
+                >
+                  <Package className="w-1 h-1" />
+                  {t("general.actions.delete")}
+                </Button>
+              )
+            : canRestoreTypeHintConfig && (
+                <Button
+                  disabled={isLoading}
+                  className={`${
+                    showEditButton ? "w-full" : "w-full"
+                  } min-h-3 bg-success-500 hover:bg-success-600 text-white-50 transition-all`}
+                  onClick={handleUnDelete}
+                >
+                  <PackageOpen className="w-1 h-1" />
+                  {t("general.actions.restore")}
+                </Button>
+              )}
 
-          {showEditButton && renderEditForm && (
+          {showEditButton && renderEditForm && canUpdateTypeHintConfig && (
             <Button
               disabled={isLoading}
               className="w-full min-h-3 bg-gray-500 hover:bg-gray-600 text-white-50 transition-all"
@@ -242,7 +285,7 @@ const TypeHintConfigCardActions = <
         </div>
       </div>
 
-      {renderEditForm && (
+      {renderEditForm && canUpdateTypeHintConfig && (
         <Modal
           isOpen={isEditModalOpen}
           onClose={() => setIsEditModalOpen(false)}
